// backend/src/database/generated/kingdomino.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: 7efTfB/1WjOMuNPVnJpoJaGayY5+EHBCyyhe0qNZurk1n1sWt7TuJjpMZBMCQ9gshx+a2RKvotQVivUboX+bnw==
 */

/* eslint-disable */
// tslint:disable

import KingdominoOptions from './kingdomino_options'

interface Kingdomino {
  deck: (Array<number | null>) | null
  drawn_dominos: (Array<number | null>) | null
  /**
   * @default nextval('kingdomino_id_seq'::regclass)
   */
  id: number & {readonly __brand?: 'kingdomino_id'}
  options: (KingdominoOptions['id']) | null
  players: (Array<number | null>) | null
}
export default Kingdomino;

interface Kingdomino_InsertParameters {
  deck?: (Array<number | null>) | null
  drawn_dominos?: (Array<number | null>) | null
  /**
   * @default nextval('kingdomino_id_seq'::regclass)
   */
  id?: number & {readonly __brand?: 'kingdomino_id'}
  options?: (KingdominoOptions['id']) | null
  players?: (Array<number | null>) | null
}
export type {Kingdomino_InsertParameters}